refactors convex folder for #82
