+++ convex/auth.ts
@@ -2,7 +2,6 @@ import { createClient, type GenericCtx } from "@convex-dev/better-auth";
 import { convex } from "@convex-dev/better-auth/plugins";
 import { components } from "./_generated/api";
 import { DataModel } from "./_generated/dataModel";
-import { query } from "./_generated/server";
 import { betterAuth } from "better-auth";
 import authConfig from "./auth.config";
 
@@ -12,27 +11,21 @@ const siteUrl = process.env.SITE_URL!;
 // as well as helper methods for general use.
 export const authComponent = createClient<DataModel>(components.betterAuth);
 
-export const createAuth = (
-  ctx: GenericCtx<DataModel>,
-) => {
+export const createAuth = (ctx: GenericCtx<DataModel>) => {
   return betterAuth({
-    // disable logging when createAuth is called just to generate options.
-    // this is not required, but there's a lot of noise in logs without it.
     baseURL: siteUrl,
     database: authComponent.adapter(ctx),
-    // Configure simple, non-verified email/password to get started
     emailAndPassword: {
       enabled: true,
       requireEmailVerification: false,
     },
     socialProviders: {
-      google: { 
-          clientId: process.env.GOOGLE_CLIENT_ID as string, 
-          clientSecret: process.env.GOOGLE_CLIENT_SECRET as string, 
-        }, 
+      google: {
+        clientId: process.env.GOOGLE_CLIENT_ID as string,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET as string,
+      },
     },
     plugins: [
-      // The Convex plugin is required for Convex compatibility
       convex({
         authConfig,
         jwksRotateOnTokenGenerationError: true,
@@ -40,12 +33,3 @@ export const createAuth = (
     ],
   });
 };
-
-// Example function for getting the current user
-// Feel free to edit, omit, etc.
-export const getCurrentUser = query({
-  args: {},
-  handler: async (ctx) => {
-    return authComponent.getAuthUser(ctx);
-  },
-});
