+++ convex/features/chat/agent.ts
@@ -1,6 +1,6 @@
 import { Agent, createTool } from "@convex-dev/agent";
-import { components } from "../_generated/api";
-import { internal } from "../_generated/api";
+import { components } from "../../_generated/api";
+import { internal } from "../../_generated/api";
 import { stepCountIs } from "ai";
 import { gateway } from "ai";
 import { z } from 'zod/v3';
@@ -14,7 +14,6 @@ export const createFlashcardTool = createTool({
     note: z.string().describe("Additional note or explanation about the flashcard"),
   }),
   handler: async (ctx, args, options): Promise<string> => {
-    // Access context provided by the agent framework
     const threadId = ctx.threadId;
     const userId = ctx.userId;
     const messageId = ctx.messageId || "pending";
@@ -23,8 +22,6 @@ export const createFlashcardTool = createTool({
       return "Missing context for creating flashcard approval.";
     }
     
-    // Access toolCallId from options (third parameter)
-    // The ToolCallOptions type has additional properties at runtime including toolCallId
     const optionsWithId = options as ToolCallOptions & { toolCallId?: string };
     const toolCallId = optionsWithId?.toolCallId;
     
@@ -33,9 +30,8 @@ export const createFlashcardTool = createTool({
       return "Error: Unable to create flashcard approval without toolCallId.";
     }
     
-    // Create the approval request (idempotent based on content)
     try {
-      await ctx.runMutation(internal.chat.flashcardApprovals.createApprovalRequestInternal, {
+      await ctx.runMutation(internal.features.chat.flashcardApprovals.createApprovalRequestInternal, {
         threadId,
         messageId,
         toolCallId,
@@ -60,7 +56,7 @@ export const agent: Agent = new Agent(components.agent, {
   You are a friendly language teacher assistant. Explain grammar and vocabulary and create spanish flashcards for the user. Do not ask for permission to create flashcards and just create them. 
   Make sure to only create flashcards for sentences not explanations or concepts. If the user asked you for a concept, create example sentences that show this concept in action. 
   `,
-  stopWhen: stepCountIs(10), // Limit tool call steps
+  stopWhen: stepCountIs(10),
   tools: {
     createFlashcard: createFlashcardTool,
   },
