+++ convex/features/chat/messages.ts
@@ -1,25 +1,22 @@
 import { v, ConvexError } from "convex/values";
-import { internalAction, mutation, query } from "../_generated/server";
+import { internalAction, mutation, query } from "../../_generated/server";
 import { paginationOptsValidator } from "convex/server";
-import { internal } from "../_generated/api";
+import { internal } from "../../_generated/api";
 import { saveMessage } from "@convex-dev/agent";
 import { listUIMessages, syncStreams } from "@convex-dev/agent";
-import { components } from "../_generated/api";
-import { authComponent } from "../auth";
+import { components } from "../../_generated/api";
+import { getAuthUser } from "../../db/users";
 import { agent } from "./agent";
-import { GoogleGenerativeAIProviderOptions } from "@ai-sdk/google";
-
 
 export type ListMessagesStreamArgs = {
   kind: "list";
   includeStatuses?: ("streaming" | "finished" | "aborted")[];
 };
 
-
 const agentComponent = components.agent;
 
 /**
- * Send a user message and trigger async AI response generation
+ * Send a user message and trigger async AI response generation.
  */
 export const sendMessage = mutation({
   args: {
@@ -28,12 +25,9 @@ export const sendMessage = mutation({
   },
   returns: v.string(),
   handler: async (ctx, args) => {
-    const user = await authComponent.getAuthUser(ctx);
-    if (!user) {
-      throw new ConvexError("Not authenticated");
-    }
+    const user = await getAuthUser(ctx);
+    if (!user) throw new ConvexError("Not authenticated");
 
-    // Verify thread ownership
     const thread = await ctx.runQuery(agentComponent.threads.getThread, {
       threadId: args.threadId,
     });
@@ -42,14 +36,12 @@ export const sendMessage = mutation({
       throw new ConvexError("Thread not found or access denied");
     }
 
-    // Save the user message
     const { messageId } = await saveMessage(ctx, agentComponent, {
       threadId: args.threadId,
       prompt: args.prompt,
     });
 
-    // Schedule async response generation
-    await ctx.scheduler.runAfter(0, internal.chat.messages.generateResponse, {
+    await ctx.scheduler.runAfter(0, internal.features.chat.messages.generateResponse, {
       threadId: args.threadId,
       promptMessageId: messageId,
     });
@@ -59,16 +51,14 @@ export const sendMessage = mutation({
 });
 
 /**
- * List messages for a thread in UI-friendly format
+ * List messages for a thread in UI-friendly format.
  */
 export const listMessages = query({
   args: {
     threadId: v.string(),
     paginationOpts: paginationOptsValidator,
-    // Stream args are permissive here to align with @convex-dev/agent validators
     streamArgs: v.optional(v.any()),
   },
-  // UIMessage includes provider-specific fields; accept any to avoid blocking streaming
   returns: v.object({
     page: v.array(v.any()),
     isDone: v.boolean(),
@@ -76,32 +66,22 @@ export const listMessages = query({
     streams: v.optional(v.any()),
   }),
   handler: async (ctx, args) => {
-    const user = await authComponent.getAuthUser(ctx);
+    const user = await getAuthUser(ctx);
     if (!user) {
-      return {
-        page: [],
-        isDone: true,
-        continueCursor: "",
-      };
+      return { page: [], isDone: true, continueCursor: "" };
     }
 
-    // Verify thread ownership
     const thread = await ctx.runQuery(agentComponent.threads.getThread, {
       threadId: args.threadId,
     });
 
     if (!thread || thread.userId !== user._id) {
-      return {
-        page: [],
-        isDone: true,
-        continueCursor: "",
-      };
+      return { page: [], isDone: true, continueCursor: "" };
     }
 
-    // Get UI messages
     const messages = await listUIMessages(ctx, agentComponent, {
       threadId: args.threadId,
-      paginationOpts: args.paginationOpts
+      paginationOpts: args.paginationOpts,
     });
 
     const streams = await syncStreams(ctx, agentComponent, {
@@ -114,7 +94,7 @@ export const listMessages = query({
 });
 
 /**
- * Generate AI response to a user message (internal action)
+ * Generate AI response to a user message (internal action).
  */
 export const generateResponse = internalAction({
   args: {
@@ -124,25 +104,11 @@ export const generateResponse = internalAction({
   returns: v.null(),
   handler: async (ctx, args) => {
     try {
-      // Stream the agent response so clients receive incremental updates
       await agent.streamText(
         ctx,
         { threadId: args.threadId },
-        {
-          promptMessageId: args.promptMessageId,
-          // providerOptions: {
-          //   google: {
-          //     thinkingConfig: {
-          //       thinkingBudget: 8192,
-          //       includeThoughts: true,
-          //     },
-          //   } satisfies GoogleGenerativeAIProviderOptions,
-          // }
-        },
-        { 
-          saveStreamDeltas: true,
-          
-        }
+        { promptMessageId: args.promptMessageId },
+        { saveStreamDeltas: true },
       );
     } catch (error) {
       console.error("Failed to generate AI response:", error);
