+++ convex/features/courses.ts
@@ -1,13 +1,15 @@
 import { v, ConvexError } from "convex/values";
-import { mutation, query } from "./_generated/server";
-import { authComponent } from "./auth";
-import { 
-  learningStyleValidator, 
-  currentLevelValidator
-} from "./types";
+import { mutation, query } from "../_generated/server";
+import { learningStyleValidator, currentLevelValidator } from "../types";
+import { getAuthUser, getUserSettings as dbGetUserSettings, getOnboardingProgress as dbGetOnboardingProgress } from "../db/users";
+import { getCoursesForUser } from "../db/courses";
+
+// ============================================================================
+// QUERIES
+// ============================================================================
 
 /**
- * Get the current user's settings
+ * Get the current user's settings.
  */
 export const getUserSettings = query({
   args: {},
@@ -24,27 +26,17 @@ export const getUserSettings = query({
   ),
   handler: async (ctx) => {
     try {
-      const user = await authComponent.getAuthUser(ctx);
-      if (!user) {
-        return null;
-      }
-
-      const userId = user._id;
-      const settings = await ctx.db
-        .query("userSettings")
-        .withIndex("by_userId", (q) => q.eq("userId", userId))
-        .first();
-
-      return settings ?? null;
+      const user = await getAuthUser(ctx);
+      if (!user) return null;
+      return (await dbGetUserSettings(ctx, user._id)) ?? null;
     } catch {
       return null;
     }
   },
 });
 
-
 /**
- * Get all courses for the authenticated user
+ * Get all courses for the authenticated user.
  */
 export const getUserCourses = query({
   args: {},
@@ -60,18 +52,9 @@ export const getUserCourses = query({
   ),
   handler: async (ctx) => {
     try {
-      const user = await authComponent.getAuthUser(ctx);
-      if (!user) {
-        return [];
-      }
-
-      const userId = user._id;
-      const courses = await ctx.db
-        .query("courses")
-        .withIndex("by_userId", (q) => q.eq("userId", userId))
-        .take(50);
-
-      return courses;
+      const user = await getAuthUser(ctx);
+      if (!user) return [];
+      return getCoursesForUser(ctx, user._id);
     } catch {
       return [];
     }
@@ -79,7 +62,7 @@ export const getUserCourses = query({
 });
 
 /**
- * Get the currently active course for the authenticated user
+ * Get the currently active course for the authenticated user.
  */
 export const getActiveCourse = query({
   args: {},
@@ -96,32 +79,20 @@ export const getActiveCourse = query({
   ),
   handler: async (ctx) => {
     try {
-      const user = await authComponent.getAuthUser(ctx);
-      if (!user) {
-        return null;
-      }
+      const user = await getAuthUser(ctx);
+      if (!user) return null;
 
-      const userId = user._id;
-      
-      // Get user settings to find active course ID
-      const settings = await ctx.db
-        .query("userSettings")
-        .withIndex("by_userId", (q) => q.eq("userId", userId))
-        .first();
-
+      const settings = await dbGetUserSettings(ctx, user._id);
       if (!settings?.activeCourseId) {
         // If no active course is set, return the first course
         const firstCourse = await ctx.db
           .query("courses")
-          .withIndex("by_userId", (q) => q.eq("userId", userId))
+          .withIndex("by_userId", (q) => q.eq("userId", user._id))
           .first();
-        
         return firstCourse;
       }
 
-      // Get the active course
-      const activeCourse = await ctx.db.get(settings.activeCourseId);
-      return activeCourse;
+      return ctx.db.get(settings.activeCourseId);
     } catch {
       return null;
     }
@@ -129,54 +100,7 @@ export const getActiveCourse = query({
 });
 
 /**
- * Set the active course for the authenticated user
- */
-export const setActiveCourse = mutation({
-  args: {
-    courseId: v.id("courses"),
-  },
-  returns: v.null(),
-  handler: async (ctx, args) => {
-    const user = await authComponent.getAuthUser(ctx);
-    if (!user) {
-      throw new ConvexError("User must be authenticated");
-    }
-
-    const userId = user._id;
-
-    // Verify the course belongs to the user
-    const course = await ctx.db.get(args.courseId);
-    if (!course) {
-      throw new ConvexError("Course not found");
-    }
-    if (course.userId !== userId) {
-      throw new ConvexError("Course does not belong to user");
-    }
-
-    // Get or create user settings
-    const existingSettings = await ctx.db
-      .query("userSettings")
-      .withIndex("by_userId", (q) => q.eq("userId", userId))
-      .first();
-
-    if (existingSettings) {
-      await ctx.db.patch(existingSettings._id, {
-        activeCourseId: args.courseId,
-      });
-    } else {
-      await ctx.db.insert("userSettings", {
-        userId,
-        hasCompletedOnboarding: true,
-        activeCourseId: args.courseId,
-      });
-    }
-
-    return null;
-  },
-});
-
-/**
- * Get the current user's onboarding progress
+ * Get the current user's onboarding progress.
  */
 export const getOnboardingProgress = query({
   args: {},
@@ -195,26 +119,52 @@ export const getOnboardingProgress = query({
   ),
   handler: async (ctx) => {
     try {
-      const user = await authComponent.getAuthUser(ctx);
-      if (!user) {
-        return null;
-      }
-
-      const userId = user._id;
-      const progress = await ctx.db
-        .query("onboardingProgress")
-        .withIndex("by_userId", (q) => q.eq("userId", userId))
-        .first();
-
-      return progress ?? null;
+      const user = await getAuthUser(ctx);
+      if (!user) return null;
+      return (await dbGetOnboardingProgress(ctx, user._id)) ?? null;
     } catch {
       return null;
     }
   },
 });
 
+// ============================================================================
+// MUTATIONS
+// ============================================================================
+
 /**
- * Save onboarding progress
+ * Set the active course for the authenticated user.
+ */
+export const setActiveCourse = mutation({
+  args: {
+    courseId: v.id("courses"),
+  },
+  returns: v.null(),
+  handler: async (ctx, args) => {
+    const user = await getAuthUser(ctx);
+    if (!user) throw new ConvexError("User must be authenticated");
+
+    const course = await ctx.db.get(args.courseId);
+    if (!course) throw new ConvexError("Course not found");
+    if (course.userId !== user._id) throw new ConvexError("Course does not belong to user");
+
+    const existingSettings = await dbGetUserSettings(ctx, user._id);
+    if (existingSettings) {
+      await ctx.db.patch(existingSettings._id, { activeCourseId: args.courseId });
+    } else {
+      await ctx.db.insert("userSettings", {
+        userId: user._id,
+        hasCompletedOnboarding: true,
+        activeCourseId: args.courseId,
+      });
+    }
+
+    return null;
+  },
+});
+
+/**
+ * Save onboarding progress.
  */
 export const saveOnboardingProgress = mutation({
   args: {
@@ -235,36 +185,23 @@ export const saveOnboardingProgress = mutation({
     baseLanguages: v.optional(v.array(v.string())),
   }),
   handler: async (ctx, args) => {
-    const user = await authComponent.getAuthUser(ctx);
-    if (!user) {
-      throw new ConvexError("User must be authenticated");
-    }
+    const user = await getAuthUser(ctx);
+    if (!user) throw new ConvexError("User must be authenticated");
 
     const userId = user._id;
 
-    // Check if progress already exists
-    const existingProgress = await ctx.db
-      .query("onboardingProgress")
-      .withIndex("by_userId", (q) => q.eq("userId", userId))
-      .first();
-
+    // Upsert onboarding progress
+    const existingProgress = await dbGetOnboardingProgress(ctx, userId);
     let progressId;
     if (existingProgress) {
       await ctx.db.patch(existingProgress._id, args);
       progressId = existingProgress._id;
     } else {
-      progressId = await ctx.db.insert("onboardingProgress", {
-        userId,
-        ...args,
-      });
+      progressId = await ctx.db.insert("onboardingProgress", { userId, ...args });
     }
 
     // Ensure userSettings exists
-    const existingSettings = await ctx.db
-      .query("userSettings")
-      .withIndex("by_userId", (q) => q.eq("userId", userId))
-      .first();
-
+    const existingSettings = await dbGetUserSettings(ctx, userId);
     if (!existingSettings) {
       await ctx.db.insert("userSettings", {
         userId,
@@ -272,22 +209,17 @@ export const saveOnboardingProgress = mutation({
         learningStyle: args.learningStyle,
       });
     } else {
-      // Update userSettings with learning preferences
-      await ctx.db.patch(existingSettings._id, {
-        learningStyle: args.learningStyle,
-      });
+      await ctx.db.patch(existingSettings._id, { learningStyle: args.learningStyle });
     }
 
     const progress = await ctx.db.get(progressId);
-    if (!progress) {
-      throw new ConvexError("Failed to retrieve onboarding progress");
-    }
+    if (!progress) throw new ConvexError("Failed to retrieve onboarding progress");
     return progress;
   },
 });
 
 /**
- * Create a new course
+ * Create a new course.
  */
 export const createCourse = mutation({
   args: {
@@ -300,22 +232,16 @@ export const createCourse = mutation({
     deckId: v.id("decks"),
   }),
   handler: async (ctx, args) => {
-    const user = await authComponent.getAuthUser(ctx);
-    if (!user) {
-      throw new ConvexError("User must be authenticated to create a course");
-    }
+    const user = await getAuthUser(ctx);
+    if (!user) throw new ConvexError("User must be authenticated to create a course");
 
-    const userId = user._id;
-
-    // Create the course with currentLevel
     const courseId = await ctx.db.insert("courses", {
       baseLanguages: args.baseLanguages,
       targetLanguages: args.targetLanguages,
       currentLevel: args.currentLevel,
-      userId,
+      userId: user._id,
     });
 
-    // Auto-create a deck for this course
     const deckName = `Learning ${args.targetLanguages.join(", ")}`;
     const deckId = await ctx.db.insert("decks", {
       courseId,
@@ -328,7 +254,7 @@ export const createCourse = mutation({
 });
 
 /**
- * Complete onboarding by creating user settings and first course
+ * Complete onboarding by creating user settings and first course.
  */
 export const completeOnboarding = mutation({
   args: {},
@@ -338,32 +264,18 @@ export const completeOnboarding = mutation({
     deckId: v.id("decks"),
   }),
   handler: async (ctx) => {
-    const user = await authComponent.getAuthUser(ctx);
-    if (!user) {
-      throw new ConvexError("User must be authenticated");
-    }
+    const user = await getAuthUser(ctx);
+    if (!user) throw new ConvexError("User must be authenticated");
 
     const userId = user._id;
 
-    // Get onboarding progress
-    const progress = await ctx.db
-      .query("onboardingProgress")
-      .withIndex("by_userId", (q) => q.eq("userId", userId))
-      .first();
+    const progress = await dbGetOnboardingProgress(ctx, userId);
+    if (!progress) throw new ConvexError("Onboarding progress not found");
 
-    if (!progress) {
-      throw new ConvexError("Onboarding progress not found");
-    }
-
-    // Get or create user settings
-    const existingSettings = await ctx.db
-      .query("userSettings")
-      .withIndex("by_userId", (q) => q.eq("userId", userId))
-      .first();
-
+    const existingSettings = await dbGetUserSettings(ctx, userId);
     const targetLanguages = progress.targetLanguages || [];
-    
-    // Create the course with data from onboarding progress
+
+    // Create the course
     const courseId = await ctx.db.insert("courses", {
       baseLanguages: progress.baseLanguages || [],
       targetLanguages,
@@ -371,7 +283,7 @@ export const completeOnboarding = mutation({
       userId,
     });
 
-    // Auto-create a deck for this course
+    // Auto-create a deck
     const deckName = `Learning ${targetLanguages.join(", ")}`;
     const deckId = await ctx.db.insert("decks", {
       courseId,
@@ -388,7 +300,6 @@ export const completeOnboarding = mutation({
         activeCourseId: courseId,
       });
     } else {
-      // Mark onboarding as complete and set active course
       await ctx.db.patch(existingSettings._id, {
         hasCompletedOnboarding: true,
         learningStyle: progress.learningStyle,
@@ -397,9 +308,10 @@ export const completeOnboarding = mutation({
       settingsId = existingSettings._id;
     }
 
-    // Delete the onboarding progress now that it's been transferred
+    // Delete the onboarding progress
     await ctx.db.delete(progress._id);
 
     return { settingsId, courseId, deckId };
   },
 });
+
